+++ backend-rust/CHANGELOG.md
@@ -8,6 +8,7 @@ Database schema version: 11
 
 ### Added
 
+- Support `include_removed` flag for `Query::bakers` query.
 - Add database migration 11 adding table tracking the removed bakers and populate the table.
 - Indexer now maintain the removed bakers table.
 - Add database migration 10 to store the `leverage_bound` and the `capital_bound` values of the chain.